+++ src/attributecode/gen.py
@@ -235,7 +235,6 @@ def load_inventory(location, from_attrib=False, base_dir=None, scancode=False, r
             running_inventory=False,
             reference_dir=reference_dir,
         )
-
         for severity, message in ld_errors:
             if 'Custom Field' in message:
                 field_name = message.replace('Custom Field: ', '').strip()