+++ src/attributecode/util.py
@@ -472,6 +472,7 @@ def ungroup_licenses(licenses):
     lic_url = []
     spdx_lic_key = []
     lic_score = []
+    lic_matched_text = []
     for lic in licenses:
         if 'key' in lic:
             lic_key.append(lic['key'])
@@ -485,7 +486,9 @@ def ungroup_licenses(licenses):
             spdx_lic_key.append(lic['spdx_license_key'])
         if 'score' in lic:
             lic_score.append(lic['score'])
-    return lic_key, lic_name, lic_file, lic_url, spdx_lic_key, lic_score
+        if 'matched_text' in lic:
+            lic_matched_text.append(lic['matched_text'])
+    return lic_key, lic_name, lic_file, lic_url, spdx_lic_key, lic_score, lic_matched_text
 
 
 # FIXME: add docstring